+++ src/pages/mypage/specialist/transaction/transactionIndex.tsx
@@ -7,7 +7,7 @@ import { useForm } from 'react-hook-form';
 
 import transactionJson from './transaction.json'
 import fetcher from '../../../utils/fetcher'
-import respVideoMeetingJson from  './resVideoMeeting.json'
+import VideoMeetingForm from './videoMeeting'
 
 interface Column {
     id: 'category' | 'title' | 'status' | 'description' | 'other';
@@ -25,21 +25,6 @@ interface Column {
     { id: 'other', label: 'その他', minWidth: 50}
   ];
 
-  function rand() {
-    return Math.round(Math.random() * 20) - 10;
-  }
-
-  function getModalStyle() {
-    const top = 50 + rand();
-    const left = 50 + rand();
-
-    return {
-      top: `${top}%`,
-      left: `${left}%`,
-      transform: `translate(-${top}%, -${left}%)`,
-    };
-  }
-
 const TransactionIndex: React.FC = () => {
     const demoData = [transactionJson, transactionJson, transactionJson]
     const history = useHistory()
@@ -49,19 +34,6 @@ const TransactionIndex: React.FC = () => {
         },
         container: {
           maxHeight: 440,
-        },
-        paper: {
-            position: 'absolute',
-            width: 400,
-            backgroundColor: theme.palette.background.paper,
-            border: '2px solid #000',
-            boxShadow: theme.shadows[5],
-            padding: theme.spacing(2, 4, 3),
-          },
-        textField: {
-            marginLeft: theme.spacing(1),
-            marginRight: theme.spacing(1),
-            width: 200,
         },
       }));
     const classes = useStyles();
@@ -77,14 +49,10 @@ const TransactionIndex: React.FC = () => {
         mode: 'onBlur',
         reValidateMode: 'onChange'
       })
-    const btnDisabler = Boolean(errors.topic) || Boolean(errors.startAt)
     type tableData = typeof initTableData[]
     type transaction = typeof transactionJson
     type transactions = transaction[]
-    type videoMeeting = typeof transactionJson.videoMeetings[0]
     type payment = typeof transactionJson.payments[0]
-    type review = typeof transactionJson.reviews[0]
-    type respVideoMeeting = typeof respVideoMeetingJson
 
     const apiUrl = "http://localhost:3000/v1";
 
@@ -101,54 +69,24 @@ const TransactionIndex: React.FC = () => {
         return res
     }
 
-    const asyncCreatePayment = async (payment: payment) => {
-        const res = await fetcher<payment>(`${apiUrl}/mypage/transaction/payment/create`, {
-            mode: 'cors',
-            method: 'POST',
-            cache: "no-cache",
-            headers: {
-                "Content-Type": "application/json",
-            },
-            body: JSON.stringify(payment)
-        })
-        return res
-    }
-
-    const asyncCreateVideoMeeting = async () => {
-        const body =  {
-            topic: videoMeeting.topic,
-            type: "2",
-            start_time: videoMeeting.start_time,
-            timezone: "Asia/Tokyo",
-            settings: {
-                use_pmi: "false"
-            }
-        }
-
-        const res = await fetcher<respVideoMeeting>(`${apiUrl}/mypage/videomeeting/create`, {
-            mode: 'cors',
-            method: 'POST',
-            cache: "no-cache",
-            headers: {
-                "Content-Type": "application/json",
-            },
-            body: JSON.stringify(body)
-        })
-        if(res.data){
-            setVideoMeeting({ ...videoMeeting, join_url: res.data.join_url})
-        }
-    }
+    // const asyncCreatePayment = async (payment: payment) => {
+    //     const res = await fetcher<payment>(`${apiUrl}/mypage/transaction/payment/create`, {
+    //         mode: 'cors',
+    //         method: 'POST',
+    //         cache: "no-cache",
+    //         headers: {
+    //             "Content-Type": "application/json",
+    //         },
+    //         body: JSON.stringify(payment)
+    //     })
+    //     return res
+    // }
 
     const [transactions, setTransactions] = useState([transactionJson])
     const [tableData, setTableData] = useState([initTableData])
-    const [status, setStatus] = useState(transactionJson.status)
-    const [review, setReview] = useState(transactionJson.reviews[0])
-    const [videoMeeting, setVideoMeeting] = useState(transactionJson.videoMeetings[0])
     const [page, setPage] = React.useState(0);
     const [rowsPerPage, setRowsPerPage] = React.useState(10);
-    const [modalStyle] = React.useState(getModalStyle);
     const [open, setOpen] = React.useState(false);
-    console.log(videoMeeting.join_url)
 
     const handleOpenModal = () => {
         setOpen(true);
@@ -206,57 +144,6 @@ const TransactionIndex: React.FC = () => {
     },[])
 
     const dummy = () => {}
-
-    const body = (
-        <div style={modalStyle} className={classes.paper}>
-          <h2>web会議を作成する</h2>
-          <p>
-            タイトルと日付を入力してください
-          </p>
-          <form className={classes.container} noValidate>
-            <TextField
-                variant="outlined"
-                margin="normal"
-                fullWidth
-                required
-                id="topic"
-                name="topic"
-                autoComplete="topic"
-                autoFocus
-                value={videoMeeting.topic}
-                onChange={(e)=>{setVideoMeeting({ ...videoMeeting, topic: e.target.value})}}
-                inputRef={register({ required: true })}
-                error={Boolean(errors.topic)}
-                helperText={errors.topic && "入力必須です"}
-            />
-            <TextField
-                id="startAt"
-                name="startAt"
-                type="datetime-local"
-                defaultValue="2017-05-24T10:30"
-                className={classes.textField}
-                InputLabelProps={{
-                shrink: true,
-                }}
-                onChange={(e)=>{setVideoMeeting({ ...videoMeeting, start_time: e.target.value})}}
-                inputRef={register({ required: true })}
-                error={Boolean(errors.startAt)}
-                helperText={errors.startAt && "入力必須です"}
-            />
-            <Button
-                    fullWidth
-                    // type="submit"
-                    variant="contained"
-                    color="primary"
-                    disabled={btnDisabler}
-                    onClick={()=>asyncCreateVideoMeeting()}
-                >
-                    bizpackを更新する
-            </Button>
-        </form>
-        {videoMeeting.join_url && (<a href={videoMeeting.join_url}>video会議はこちら</a>)}
-        </div>
-    );
 
     return (
         <>
@@ -330,13 +217,10 @@ const TransactionIndex: React.FC = () => {
             <Modal
                 open={open}
                 onClose={handleCloseModal}
-                aria-labelledby="simple-modal-title"
-                aria-describedby="simple-modal-description"
             >
-                {body}
+                <VideoMeetingForm />
             </Modal>
         </>
-        
       );
 }
